--- a/hw10_program_optimization/new_bench.txt
+++ b/hw10_program_optimization/new_bench.txt
@@ -1,0 +1,7 @@
+goos: darwin
+goarch: amd64
+pkg: github.com/fixme_my_friend/hw10_program_optimization
+cpu: Intel(R) Core(TM) i5-1038NG7 CPU @ 2.00GHz
+BenchmarkGetDomainStat-8   	 2224167	      5402 ns/op	    7546 B/op	      64 allocs/op
+PASS
+ok  	github.com/fixme_my_friend/hw10_program_optimization	17.667s