@vlyagusha 2022-02-20
vlyagusha@MBP-Vyacheslav hw10_program_optimization % benchstat new_bench.txt old_bench.txt         
name             old time/op    new time/op       delta
GetDomainStat-8    5.40µs ± 0%  229639.26µs ± 0%   ~     (p=1.000 n=1+1)

name             old alloc/op   new alloc/op      delta
GetDomainStat-8    7.55kB ± 0%  132846.64kB ± 0%   ~     (p=1.000 n=1+1)

name             old allocs/op  new allocs/op     delta
GetDomainStat-8      64.0 ± 0%    1700096.0 ± 0%   ~     (p=1.000 n=1+1)