@vlyagusha 2022-02-20
Чтобы было более показательно, сравнивал бенчи в обе стороны, то есть 
benchstat old_bench.txt new_bench.txt  
и
benchstat new_bench.txt old_bench.txt

Дополнительно запушил результаты бенчей, но наверное это избыточно.
Также не стал удалять старую реализацию функции, кмк так и ревьюить, и бенчить проще.